# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NCDMB API Documentation",
        "description": "Comprehensive API documentation for the NCDMB Document Management System. This API provides endpoints for document management, workflow automation, procurement, legal processes, and more.",
        "contact": {
            "email": "support@ncdmb.gov.ng"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "List all projects",
                "operationId": "8d4a13ef3fc773ff4031e515460919ea",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Create a new project",
                "operationId": "a4dd4c17bf5fc24c224357f03dcc848a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProjectRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Project created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/projects/{id}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get a specific project",
                "operationId": "af32aed47c19d517816e0b970d60cd5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Projects"
                ],
                "summary": "Update a project",
                "operationId": "491d11cc6efd20335421a859159e2487",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProjectRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Project updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Projects"
                ],
                "summary": "Delete a project",
                "operationId": "187865e30ff6b72f1cea3037d9dd84fa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/procurement/committees": {
            "get": {
                "tags": [
                    "Procurement"
                ],
                "summary": "List all evaluation committees",
                "operationId": "d62548672e3233c7ab6f6a30ec62ef6f",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Procurement"
                ],
                "summary": "Create a new evaluation committee",
                "operationId": "a2da589cc54e49012a0bdb7af0dffcbd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProjectEvaluationCommitteeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Committee created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/procurement/committees/{id}": {
            "get": {
                "tags": [
                    "Procurement"
                ],
                "summary": "Get a specific evaluation committee",
                "operationId": "e8ee978a6ff4990ad58e29e515fad649",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Procurement"
                ],
                "summary": "Update an evaluation committee",
                "operationId": "63f1d844a30d1db007a2658394cbe71f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProjectEvaluationCommitteeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Committee updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Procurement"
                ],
                "summary": "Delete an evaluation committee",
                "operationId": "a903d0d35589b276d0206400e0b82e27",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Committee deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/procurement/committees/{id}/dissolve": {
            "post": {
                "tags": [
                    "Procurement"
                ],
                "summary": "Dissolve an evaluation committee",
                "operationId": "83bed71e6e3bc813f63c4408a43e9597",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Committee dissolved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Success": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "data": {
                        "type": "object"
                    },
                    "message": {
                        "type": "string",
                        "example": "Operation successful"
                    }
                },
                "type": "object"
            },
...
            "ProjectRequest": {
                "required": [
                    "user_id",
                    "department_id",
                    "threshold_id",
                    "project_category_id",
                    "title",
                    "type",
                    "status",
                    "project_type",
                    "priority"
                ],
                "properties": {
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "department_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "threshold_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "project_category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "maxLength": 500,
                        "example": "Infrastructure Development Project"
                    },
                    "description": {
                        "type": "string",
                        "example": "Detailed project description",
                        "nullable": true
                    },
                    "total_proposed_amount": {
                        "type": "number",
                        "minimum": 0,
                        "example": 1000000,
                        "nullable": true
                    },
                    "sub_total_amount": {
                        "type": "number",
                        "minimum": 0,
                        "example": 900000,
                        "nullable": true
                    },
                    "service_charge_percentage": {
                        "type": "integer",
                        "maximum": 100,
                        "minimum": 0,
                        "example": 5,
                        "nullable": true
                    },
                    "markup_amount": {
                        "type": "number",
                        "minimum": 0,
                        "example": 50000,
                        "nullable": true
                    },
                    "vat_amount": {
                        "type": "number",
                        "minimum": 0,
                        "example": 50000,
                        "nullable": true
                    },
                    "proposed_start_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-15",
                        "nullable": true
                    },
                    "proposed_end_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-12-31",
                        "nullable": true
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "staff",
                            "third-party"
                        ],
                        "example": "staff"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "registered",
                            "approved",
                            "denied",
                            "kiv",
                            "discussed"
                        ],
                        "example": "pending"
                    },
                    "project_type": {
                        "type": "string",
                        "enum": [
                            "capital",
                            "operational",
                            "maintenance",
                            "research",
                            "infrastructure"
                        ],
                        "example": "infrastructure"
                    },
                    "priority": {
                        "type": "string",
                        "enum": [
                            "critical",
                            "high",
                            "medium",
                            "low"
                        ],
                        "example": "high"
                    },
                    "strategic_alignment": {
                        "type": "string",
                        "maxLength": 1000,
                        "example": "Strategic alignment notes",
                        "nullable": true
                    },
                    "fund_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "budget_year": {
                        "type": "string",
                        "maxLength": 20,
                        "example": "2024",
                        "nullable": true
                    },
                    "variation_amount": {
                        "type": "number",
                        "example": 0,
                        "nullable": true
                    },
                    "total_approved_amount": {
                        "type": "number",
                        "minimum": 0,
                        "example": 1000000,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ProjectEvaluationCommitteeRequest": {
                "required": [
                    "project_id",
                    "committee_name",
                    "committee_type",
                    "chairman_id"
                ],
                "properties": {
                    "project_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "committee_name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Technical Evaluation Committee"
                    },
                    "committee_type": {
                        "type": "string",
                        "enum": [
                            "tender_board",
                            "technical",
                            "financial",
                            "opening"
                        ],
                        "example": "technical"
                    },
                    "chairman_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "members": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "user_id": {
                                    "type": "integer",
                                    "example": 2
                                },
                                "role": {
                                    "type": "string",
                                    "enum": [
                                        "chairman",
                                        "secretary",
                                        "member",
                                        "observer"
                                    ],
                                    "example": "member"
                                }
                            },
                            "type": "object"
                        },
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "dissolved"
                        ],
                        "example": "active",
                        "nullable": true
                    },
                    "formed_at": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-15",
                        "nullable": true
                    },
                    "dissolved_at": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-12-31",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum Bearer Token Authentication. Enter token in format: Bearer <token>",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}